Add check for installed features using feature repo url

+++ ga/22.0.0.12/kernel/helpers/build/configure.sh
@@ -133,14 +133,13 @@ function main() {
     parseProviders $SEC_SSO_PROVIDERS
   fi
 
-  if [ "$SKIP_FEATURE_INSTALL" != "true" ]; then
+  if [ "$SKIP_FEATURE_INSTALL" != "true" ] && [ "$FEATURES_INSTALLED" == "false" ] ; then
     # Install needed features
     if [ "$FEATURE_REPO_URL" ]; then
       curl -k --fail $FEATURE_REPO_URL > /tmp/repo.zip
       installUtility install --acceptLicense defaultServer --from=/tmp/repo.zip || rc=$?; if [ $rc -ne 22 ]; then exit $rc; fi
       rm -rf /tmp/repo.zip
-    # Otherwise, if features.sh did not run, install server features.
-    elif [ "$FEATURES_INSTALLED" == "false" ]; then
+    else
       installUtility install --acceptLicense defaultServer || rc=$?; if [ $rc -ne 22 ]; then exit $rc; fi
     fi
   fi
